requirements.txt
Django>=5.2,<6
djangorestframework>=3.15
requests>=2.31
vosk>=0.3.45
psycopg2-binary>=2.9
